+++ sbin/ucserv.py
@@ -85,10 +85,10 @@ eva.logs.start()
 
 eva.traphandler.update_config(cfg)
 eva.udpapi.update_config(cfg)
+eva.api.update_config(cfg)
 eva.upnp.update_config(cfg)
 eva.upnp.port = 1911
 eva.lora.update_config(cfg)
-eva.api.update_config(cfg)
 eva.sysapi.update_config(cfg)
 eva.uc.modbus.update_config(cfg)