+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ Common
    * MessagePack support, internal API calls changed to msgpack
    * MQTT inter-connect API encryption strength increased to AES256
    * Core workers are now async to improve system performance
+   * Controller auto-discovery in local network via UPnP
 
 UC
 --